TimingSolver: moved to public include dir
